+++ src/App.js
@@ -1,4 +1,3 @@
-import logo from './logo.svg';
 import React from 'react'
 
 export default function App() {
